requirements.txt
asgiref==3.5.2
dj-database-url==1.2.0
Django==4.1.4
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-formtools==2.4
django-grappelli==3.0.4
django-inline-actions
django-jet==1.0.8
django-suit==0.2.28
gunicorn==20.1.0
httplib2==0.21.0
pytz==2022.7
sqlparse==0.4.3
whitenoise==6.2.0
yapf==0.32.0
psycopg2-binary==2.9.5
djangorestframework
drf-yasg
markdown
django-filter
python-dotenv==0.21.0
django-cities==0.6.1
django-cors-headers==3.13.0
ipdb==0.13.11
factory_boy
pip-tools==6.12.1
pip-reqs==0.11.0
pylint==2.15.9
pytest==7.2.0
pytest-django
pytest-cov
pytest-pythonpath
mock
pre-commit
schema==0.7.5
pyparsing
pydot
pydotplus
pillow==9.3.0
djangorestframework-simplejwt==5.2.2
graphviz
django-jazzmin
django-environ==0.4.5
google-cloud-secret-manager==2.6.0
django-storages[google]==1.11.1
requests
six
appdirs
